+++ src/main/java/net/more_rpg_classes/mixin/LivingEntityMixin.java
@@ -17,19 +17,18 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(LivingEntity.class)
 public abstract class LivingEntityMixin {
 
-    @Inject(method = "createLivingAttributes()Lnet/minecraft/entity/attribute/DefaultAttributeContainer$Builder;",
-            require = 1, allow = 1, at = @At("RETURN"))
-    private static void moreEntityAttributes$addAttributes
-            (CallbackInfoReturnable<DefaultAttributeContainer.Builder> info){
-        info.getReturnValue().add(MRPGCEntityAttributes.DAMAGE_REFLECT_MODIFIER);
-        info.getReturnValue().add(MRPGCEntityAttributes.ARCANE_FUSE_MODIFIER);
-        info.getReturnValue().add(MRPGCEntityAttributes.LIFESTEAL_MODIFIER);
-        info.getReturnValue().add(MRPGCEntityAttributes.RAGE_MODIFIER);
+    @Inject(method = "createLivingAttributes", at = @At("RETURN"))
+    private static void mrpgc_lib$createLivingAttributes(CallbackInfoReturnable<DefaultAttributeContainer.Builder> cir) {
+        cir.getReturnValue()
+                .add(MRPGCEntityAttributes.ARCANE_FUSE_MODIFIER)
+                .add(MRPGCEntityAttributes.DAMAGE_REFLECT_MODIFIER)
+                .add(MRPGCEntityAttributes.LIFESTEAL_MODIFIER)
+                .add(MRPGCEntityAttributes.RAGE_MODIFIER)
+        ;
     }
 
-
     @Inject(method = "damage", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/LivingEntity;applyDamage(Lnet/minecraft/entity/damage/DamageSource;F)V"))
-    private void applyBeforeDamage(DamageSource source, float amount, CallbackInfoReturnable<Boolean> cir) {
+    private void damageReflect$damage(DamageSource source, float amount, CallbackInfoReturnable<Boolean> cir) {
         EntityAttributeInstance dmgReflect = ((LivingEntity) (Object) this)
                 .getAttributeInstance(MRPGCEntityAttributes.DAMAGE_REFLECT_MODIFIER);
         int value1 = (int) dmgReflect.getValue();